Moves api functions to routes/index.js; updates port number
Moved the api router functions and requests to its own file.
Added new routes for '/vadio' which routes to the search page and
'/searching' which routes the ajax request to the server when
the form is submitted. Added a getVadio() function which handles the
external API request and returns the response back to the client/ajax.